+++ rosomaxa/src/algorithms/gsom/contraction.rs
@@ -32,6 +32,10 @@ where
         .filter(|coord| coord.0 % x_decim == 0 || coord.1 % y_decim == 0)
         .collect::<Vec<_>>();
 
+    if network.size() - removed.len() < 4 {
+        return;
+    }
+
     // remove nodes with given coordinates, but keep track of their data
     let data = removed.iter().fold(Vec::new(), |mut data, coordinate| {
         let node = network.get_mut(coordinate).unwrap();
@@ -51,23 +55,6 @@ where
 
         node
     });
-
-    // NOTE: this is unfortunate, probably, compact was called too often on low number of nodes
-    if network.size() == 0 {
-        let dimension = network.dimension();
-        let get_weights = |idx: usize| -> Vec<Float> {
-            data.get(idx).map_or_else(
-                // NOTE: probably, very bad initialization..
-                || (0..dimension).map(|_| idx as Float).collect(),
-                |data| data.weights().to_vec(),
-            )
-        };
-
-        network.insert(context, (0, 0).into(), get_weights(0).as_slice());
-        network.insert(context, (0, 1).into(), get_weights(1).as_slice());
-        network.insert(context, (1, 0).into(), get_weights(2).as_slice());
-        network.insert(context, (1, 1).into(), get_weights(3).as_slice());
-    }
 
     // reintroduce data from deleted notes to the network while network growth is not allowed.
     // NOTE: min_max_weights is not changed here