+++ vrp-cli/tests/unit/commands/analyze_test.rs
@@ -1,23 +1,13 @@
 use super::*;
+use crate::cli::{get_app, run_subcommand};
 
 const PRAGMATIC_PROBLEM_PATH: &str = "../examples/data/pragmatic/simple.basic.problem.json";
-
-struct DummyWrite {}
-
-impl Write for DummyWrite {
-    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-        Ok(buf.len())
-    }
-
-    fn flush(&mut self) -> std::io::Result<()> {
-        Ok(())
-    }
-}
 
 #[test]
 fn can_run_analyze_dbscan() {
     let tmpfile = tempfile::NamedTempFile::new().unwrap();
     let args = vec![
+        "vrp-cli",
         "analyze",
         "dbscan",
         "pragmatic",
@@ -25,9 +15,9 @@ fn can_run_analyze_dbscan() {
         "--out-result",
         tmpfile.path().to_str().unwrap(),
     ];
-    let matches = get_analyze_app().try_get_matches_from(args).unwrap();
+    let matches = get_app().try_get_matches_from(args).unwrap();
 
-    run_analyze(&matches, |_| BufWriter::new(Box::new(DummyWrite {}))).unwrap();
+    run_subcommand(matches);
 }
 
 #[test]
@@ -41,6 +31,7 @@ fn can_detect_wrong_argument_in_dbscan() {
 fn can_run_analyze_kmedoids() {
     let tmpfile = tempfile::NamedTempFile::new().unwrap();
     let args = vec![
+        "vrp-cli",
         "analyze",
         "kmedoids",
         "pragmatic",
@@ -50,7 +41,7 @@ fn can_run_analyze_kmedoids() {
         "--out-result",
         tmpfile.path().to_str().unwrap(),
     ];
-    let matches = get_analyze_app().try_get_matches_from(args).unwrap();
+    let matches = get_app().try_get_matches_from(args).unwrap();
 
-    run_analyze(&matches, |_| BufWriter::new(Box::new(DummyWrite {}))).unwrap();
+    run_subcommand(matches);
 }