+++ vrp-cli/tests/unit/commands/check_test.rs
@@ -1,4 +1,4 @@
-use super::*;
+use crate::cli::{get_app, run_subcommand};
 
 const PRAGMATIC_PROBLEM_PATH: &str = "../examples/data/pragmatic/simple.basic.problem.json";
 const PRAGMATIC_MATRIX_PATH: &str = "../examples/data/pragmatic/simple.basic.matrix.json";
@@ -7,6 +7,7 @@ const PRAGMATIC_SOLUTION_PATH: &str = "../examples/data/pragmatic/simple.basic.s
 #[test]
 fn can_run_check_solution() {
     let args = vec![
+        "vrp-cli",
         "check",
         "pragmatic",
         "--problem-file",
@@ -16,7 +17,7 @@ fn can_run_check_solution() {
         "--solution-file",
         PRAGMATIC_SOLUTION_PATH,
     ];
-    let matches = get_check_app().try_get_matches_from(args).unwrap();
+    let matches = get_app().try_get_matches_from(args).unwrap();
 
-    run_check(&matches).unwrap();
+    run_subcommand(matches);
 }