+++ vrp-cli/tests/unit/commands/generate_test.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::cli::{get_app, run_subcommand};
 use vrp_cli::pragmatic::format::problem::PragmaticProblem;
 
 const PRAGMATIC_PROBLEM_PATH: &str = "../examples/data/pragmatic/simple.basic.problem.json";
@@ -7,6 +8,7 @@ const PRAGMATIC_PROBLEM_PATH: &str = "../examples/data/pragmatic/simple.basic.pr
 fn can_generate_problem_from_args() {
     let tmpfile = tempfile::NamedTempFile::new().unwrap();
     let args = vec![
+        "vrp-cli",
         "generate",
         "pragmatic",
         "--prototypes",
@@ -18,9 +20,9 @@ fn can_generate_problem_from_args() {
         "--out-result",
         tmpfile.path().to_str().unwrap(),
     ];
-    let matches = get_generate_app().try_get_matches_from(args).unwrap();
+    let matches = get_app().try_get_matches_from(args).unwrap();
 
-    run_generate(&matches).unwrap();
+    run_subcommand(matches);
 
     let problem = BufReader::new(tmpfile.as_file()).read_pragmatic().unwrap();
     assert_eq!(problem.jobs.size(), 100);